# docs/swagger.yaml
basePath: /api
definitions:
...
  entity.ProjectRecruitmentType:
    enum:
    - MT_Management Trainee
    - PH_Professional Hire
    - NS_Non Staff to Staff
    type: string
    x-enum-varnames:
    - PROJECT_RECRUITMENT_TYPE_MT
    - PROJECT_RECRUITMENT_TYPE_PH
    - PROJECT_RECRUITMENT_TYPE_NS
...
  entity.TemplateActivityLineStatus:
    enum:
    - ACTIVE
    - INACTIVE
    type: string
    x-enum-varnames:
    - TEMPLATE_ACTIVITY_LINE_STATUS_ACTIVE
    - TEMPLATE_ACTIVITY_LINE_STATUS_INACTIVE
  entity.TemplateActivityStatus:
    enum:
    - ACTIVE
    - INACTIVE
    type: string
    x-enum-varnames:
    - TEMPLATE_ACTIVITY_STATUS_ACTIVE
    - TEMPLATE_ACTIVITY_STATUS_INACTIVE
...
  entity.TemplateQuestionFormType:
    enum:
    - ADMINISTRATIVE_SELECTION
    - TEST
    - INTERVIEW
    - FGD
    - FINAL_INTERVIEW
    - OFFERING_LETTER
    - CONTRACT_DOCUMENT
    - DOCUMENT_CHECKING
    - SURAT_IZIN_ORANG_TUA
    - KARYAWAN_TETAP
    - PKWT
    - PKWTT
    type: string
    x-enum-varnames:
    - TQ_FORM_TYPE_ADMINISTRATIVE_SELECTION
    - TQ_FORM_TYPE_TEST
    - TQ_FORM_TYPE_INTERVIEW
    - TQ_FORM_TYPE_FGD
    - TQ_FORM_TYPE_FINAL_INTERVIEW
    - TQ_FORM_TYPE_OFFERING_LETTER
    - TQ_FORM_TYPE_CONTRACT_DOCUMENT
    - TQ_FORM_TYPE_DOCUMENT_CHECKING
    - TQ_FORM_TYPE_SURAT_IZIN_ORTU
    - TQ_FORM_TYPE_KARYAWAN_TETAP
    - TQ_FORM_TYPE_PKWT
    - TQ_FORM_TYPE_PKWTT
  entity.TemplateQuestionStatus:
    enum:
    - ACTIVE
    - INACTIVE
    type: string
    x-enum-varnames:
    - TEMPLATE_QUESTION_STATUS_ACTIVE
    - TEMPLATE_QUESTION_STATUS_INACTIVE
...
  request.CreateOrUpdateTemplateActivityLineRequest:
    properties:
      deleted_template_activity_line_ids:
        items:
          type: string
        type: array
      template_activity_id:
        type: string
      template_activity_lines:
        items:
          properties:
            color_hex_code:
              type: string
            description:
              type: string
            id:
              type: string
            name:
              type: string
            status:
              type: string
            template_question_id:
              type: string
          required:
          - name
          - status
          - template_question_id
          type: object
        type: array
    required:
    - template_activity_id
    - template_activity_lines
    type: object
...
  request.CreateTemplateActivityRequest:
    properties:
      description:
        type: string
      name:
        type: string
      recruitment_type:
        type: string
      status:
        type: string
    required:
    - name
    - recruitment_type
    - status
    type: object
...
  request.UpdateTemplateActivityRequest:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      recruitment_type:
        type: string
      status:
        type: string
    required:
    - id
    - name
    - recruitment_type
    - status
    type: object
...
  response.AnswerTypeResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  response.DocumentSetupResponse:
    properties:
      body:
        type: string
      created_at:
        type: string
      document_type:
        $ref: '#/definitions/response.DocumentTypeResponse'
      document_type_id:
        type: string
      footer:
        type: string
      header:
        type: string
      id:
        type: string
      recruitment_type:
        $ref: '#/definitions/entity.ProjectRecruitmentType'
      title:
        type: string
      updated_at:
        type: string
    type: object
  response.DocumentTypeResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  response.DocumentVerificationResponse:
    properties:
      format:
        type: string
      id:
        type: string
      name:
        type: string
      template_question:
        $ref: '#/definitions/response.TemplateQuestionResponse'
      template_question_id:
        type: string
    type: object
...
  response.QuestionOptionResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      option_text:
        type: string
      question_id:
        type: string
      updated_at:
        type: string
    type: object
  response.QuestionResponse:
    properties:
      answer_type_id:
        type: string
      answer_types:
        $ref: '#/definitions/response.AnswerTypeResponse'
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      question_options:
        items:
          $ref: '#/definitions/response.QuestionOptionResponse'
        type: array
      template_question_id:
        type: string
      updated_at:
        type: string
    type: object
...
  response.TemplateActivityLineResponse:
    properties:
      color_hex_code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      question_template_id:
        type: string
      status:
        $ref: '#/definitions/entity.TemplateActivityLineStatus'
      template_activity:
        $ref: '#/definitions/response.TemplateActivityResponse'
      template_activity_id:
        type: string
      template_question:
        $ref: '#/definitions/response.TemplateQuestionResponse'
      updated_at:
        type: string
    type: object
  response.TemplateActivityResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      recruitment_type:
        $ref: '#/definitions/entity.ProjectRecruitmentType'
      status:
        $ref: '#/definitions/entity.TemplateActivityStatus'
      template_activity_lines:
        items:
          $ref: '#/definitions/response.TemplateActivityLineResponse'
        type: array
      updated_at:
        type: string
    type: object
  response.TemplateQuestionResponse:
    properties:
      description:
        type: string
      document_setup:
        $ref: '#/definitions/response.DocumentSetupResponse'
      document_setup_id:
        type: string
      document_verifications:
        items:
          $ref: '#/definitions/response.DocumentVerificationResponse'
        type: array
      duration:
        type: integer
      form_type:
        $ref: '#/definitions/entity.TemplateQuestionFormType'
      id:
        type: string
      name:
        type: string
      questions:
        items:
          $ref: '#/definitions/response.QuestionResponse'
        type: array
      status:
        $ref: '#/definitions/entity.TemplateQuestionStatus'
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Julong Recruitment API Docs
  version: "1.0"
paths:
...
  /template-activities:
    get:
      consumes:
      - application/json
      description: Find all template activities paginated
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: page_size
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: Created At
        in: query
        name: created_at
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TemplateActivityResponse'
      security:
      - BearerAuth: []
      summary: Find all template activities paginated
      tags:
      - Template Activities
    post:
      consumes:
      - application/json
      description: Create template activity
      parameters:
      - description: Create Template Activity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateTemplateActivityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.TemplateActivityResponse'
      security:
      - BearerAuth: []
      summary: Create template activity
      tags:
      - Template Activities
  /template-activities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete template activity
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete template activity
      tags:
      - Template Activities
    get:
      consumes:
      - application/json
      description: Find template activity by id
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TemplateActivityResponse'
      security:
      - BearerAuth: []
      summary: Find template activity by id
      tags:
      - Template Activities
  /template-activities/update:
    put:
      consumes:
      - application/json
      description: Update template activity
      parameters:
      - description: Update Template Activity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateTemplateActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TemplateActivityResponse'
      security:
      - BearerAuth: []
      summary: Update template activity
      tags:
      - Template Activities
  /template-activity-lines:
    post:
      consumes:
      - application/json
      description: Create or update template activity line
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.CreateOrUpdateTemplateActivityLineRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.TemplateActivityResponse'
      security:
      - BearerAuth: []
      summary: Create or update template activity line
      tags:
      - Template Activity Lines
  /template-activity-lines/template-activity/{id}:
    get:
      consumes:
      - application/json
      description: Find template activity line by template activity id
      parameters:
      - description: Template Activity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TemplateActivityLineResponse'
      security:
      - BearerAuth: []
      summary: Find template activity line by template activity id
      tags:
      - Template Activity Lines
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"